Update 1inch unit test due to recent changes
Slippage `require`s make mocking a bit more tricky now.  We
updated the mock so that the slippage calculation can be passed
in, so as to keep the mock setup as self-contained as possible.

We had to then change the assert from checking equality with
amounts to checking that the received amount satisfies reasonable
constraints, given the slippage.

We also have to send ether to the manager contract before the swap,
as the swap function checks for the zero address (1inch convention
is to use the zero address to represent ETH), and then attempts to
send value with the swap call to 1inch.