Add rebalancing functionality to manager
Manager has added references to a strategy and a
liquidity pool.  Interfaces were created for
later extensibility.

The pool was extended to allow it to be drained
by the manager, which then uses the ETH to swap
for assets in the rebalance function.

Base strategy takes in and validates input asset data.

The rebalance function cycles through input assets,
which are assumed to be given as pairs of ERC20 and
whole number percentages.  For each asset, the appropriate
amount of the ETH drained from the pool is then
swapped using the 1inch swap function.

Unit tests were added to test the strategy contract
and rebalance function.