Complete the DAI3 contract (basic reborrowing)
Reinvest COMP into DAI and be able to close out the cDAI
position at the end.  This was already done for the
LeveragedYieldFarm contract but had been left undone here.

To enable swap functionality on the contract, a 1inch
swap base contract was created.

The purpose of the DAI3 contract for now is just for testing.
The flash loan version of the strategy (in the contract
LeveragedYieldFarm) will be adapted to follow a similar
structure with the same base contracts and will probably
be the one included in the APYManager.