Restructure Manager and Strategy interaction
The fundamental change is that the strategy contract owns
the funds.  This is to allow it to reinvest farmed assets
and possibly other byproducts into the strategy; otherwise,
there would be transfers back-and-forth between the two
contracts.

The manager now kicks off the different stages of the strategy,
by invoking its functions, and sends it ETH as needed.  The
manager doesn't need to know what the input assets are, etc.

This should extend nicely to when the strategy contract is the
strategy executor.