Fix upgrade initialize issue
Initializing the upgrades is a little tricky, due to the way `Initializable` works, there is only one private var used to check the `initializer` modifier.  This means you can't call the same initialize function in the next version of the implementation!

Not only that, if you create a new initialize function for v2, you can't use the `initializer` modifier anymore.  See this issue that's been open since Feb 2019:
https://github.com/OpenZeppelin/openzeppelin-sdk/issues/637

The way I've gotten around this is to set an admin on the pool impl contract (this should be the same as the proxy admin).  The first initialize already set the deployer of the impl contract as the owner.  So now only the owner can set the admin and only the admin can invoke the subsequent initializes.