Replace drain functionality with approval
In line with recent discussions on separation of concerns
of the manager and the pools, the drain function has been
removed.

The new workflow is that each pool must approve the manager
to transfer funds out.  This decoupling of manager and pools
also offers better security.

A pool has functions to allow its owner to do an infinite
approval or revoke it.  The revoke function is not controlled
by the Pausable mechanism, so as to allow it to function when
the pool is locked.