Deploy accounts and handle all tokens in factory
Renaming suggests the overall structure of work.  The factory
is the deployer of funds, and each deployed contract is simply
an account to hold the funds, with sufficient logic to allow
withdrawals.

The factory/deployer still needs to handle some token registration
as the chainlink registry will need to know the balances of all
used tokens.  Several functions have been created for this
functionality, but are currently not optimal, just suggestions for
future work.