Finally, all unit tests fixed with some caveats
Waffle's "calledOnContract" does not work with Hardhat.
This is a known issue with resolution in the works.
For now, the assert is commented out, and we *do* have test
coverage in the integration tests, but obviously not ideal.

Alternative is to just patch it up for the time being with the
old Gnosis mock approach.  This might also raise the question
of why not use the Gnosis mock, which is a valid question, but
the Waffle mock does, in our use-cases at least, read more simply
and cleanly for those not familiar with the codebase IMO.

Waffle is designed to work with Ethers, while Gnosis is not, so
I think fundamentally the shift is warranted.

Finally, much cleaning up still remains to do.  The tests are
crufty, with some redundancies and it is unclear at points what
we are attempting to test.