Refactor APT contract into v1 and v2 versions
In order to be rigorous in our testing, we should setup the
deploys to reflect that we will be upgrading our logic
contract for the already deployed proxy on Mainnet.

This means that in our unit testing, we should deploy the original
contract (v1) as before and then upgrade to v2.  This will make
sure there isn't a collision in storage or another unexpected
interaction.

We need to mimic actual deployment in our unit tests because
we have structured our tests so that much of the verification
of deployment setup correctness is being done in the unit testing.
The integration testing is mainly to test against Mainnet tokens
and walk through some functional requirements.

Additionally, in integration testing, we *could* point to the
already deployed proxy on Mainnet and upgrade it using the new
v2 logic contract; however this complicates test setup as the
existing pool balances aren't necessarily ideal for testing.
So for now, I am doing the same 2-stage deploy as with unit
tests.

Refactoring the contract into two versions supports the above,
but ultimately, this was the envisioned upgrade pattern when I created
the `initializeUpgrade` function.  It seems best to stick with that
approach as it is designed to be more dummy-proof and enforces
consistency in upgrades.  I basically forgot about all this when
upgrading APT for the alpha demo, but we should go back to this.