Pin exact versions of OZ packages.
We need to make sure even minor changes aren't incorporated into
our codebase, as we use an upgradeable proxy pattern and so
need to be certain storage changes did not occur.

The `openzeppelin-contracts` entry in package.json was updated
since the audit from rc-3.2.0 to 3.2.0, but the only changes
were to the release notes and version number.

The previous caret versioning means we could have pulled in newer
minor versions accidentally when deploying, but the minor updates
for both OZ packages happened in November, a couple months after
our audit and deployment, so we are safe.