Update yarn.lock
Lockfile has to be updated for the pinned versions.

Unfortunately, the previous lockfile shows we pulled in the newer
`openzeppelin-contracts` package (3.3.0), so I had to go through
the Etherscan verified code to see if any of the newer deployments
had the new contracts.  I also double-checked the old contracts
on Etherscan also.  The versions are easily distinguished by the
pragmas that include 0.7.

Summary:
- pools and APY token are safe: both packages use the pinned version
- address registry and manager:
  - logic contracts are safe
  - proxy contracts use the newer 3.3.0 package, but there are no
    storage changes.  The new changes for the proxy-related contracts
    are either renaming or adjusting unused functions in the Address
    library.