Fix issue with TVL-supply mismatch
The issue was that on each iteration of the loop, the TVL
would remain static (since it is read from a Chainlink oracle),
but the total supply of mAPT would be increasing.

This made the mAPT share calculation incorrect.  However, the fix
takes advantage of the fact that the share computation (but not
minting) will give the correct amounts, as long as minting is all
done at the end.