Fix various compilation issues (see body)
Assembly fix was straightforward.  A bigger problem is the use
of the OZ enumerable set for bytes32.  This is not in the package
until 3.3.0 and we are pinned on 3.0.0.  Rather than go through
an extensive storage check of the new contracts (seems like it's
only one change to one contract that shouldn't matter, but ...)
I thought it's safer to just copy over the relevant library.

I diffed the file to make sure it got copied over correctly; only
changed the pragma at the top, added the solhint-disable, and
changed the naming from `EnumerableSetUpgradeable` to plain
`EnumerableSet`.

Remaining issues were largely superficial except for the fact
that bytes32 isn't convertible into address. For now, I changed
everything to bytes32.