Test manual submission logic for TVL
Expiry check was adjusted to follow the same logic as for
the contract locking, i.e. the period number gives the number
of blocks for which the condition holds, including the current
block.

Test is not ideal, as it relies on intimate knowledge of the
time logic.  It would be better to have a view function that
could be queried to tell you if you are in manual submission mode
and that same function is used in `getTvl`.