Revamp testing framework to handle LUSD staking
A fair amount of cleanup in addition to some rethinking of how
to do interface overrides.  "Explicit is better than implicit" so
now the various stableswap interfaces require explicit overrides,
although "IStableSwap" remains as an alias for the most common
3-underlyer type.

The trickiest part of the change was to override specific functions,
e.g. the synthetix-style staking has a "stake" function rather than
the normal liquidity gauge "deposit".  We can do this easily in JS
since we can map "deposit(uint256)" to "stake(uint256)".  However,
this necessitated not doing a "connect" on the contract after the
mapping is added.  Instead, we explicitly set the signer in each
`getContractAt`.