Renaming PoolManager test file
For now, it seems better to keep the mAPT segregated into the two
main testing categories (agg integration with calcs and funding/
withdrawing tests with pools), as they seem orthogonal and require
different setups.  Some thought is required to better organize
these tests and perhaps we need to rethink how the "integration"
tests are segregated in general.  "Per-contract" isn't a great
division where "integration" is concerned.