Upgrade Address Registry in deployment
Since we have the new aliases in the registry being used in multiple places,
we need to upgrade the Address Registry also.

During testing, I found it was viable to deploy mocks whose addresses
match those on Mainnet.  Since the address registry is so critical,
it makes sense to hard-code that rather than pass into the constructor.