Remove proxy contract and restructure tests
We no longer need the MetaPoolTokenProxy.  In addition, the tests
for it were largely testing only OpenZeppelin code paths, none of
our logic was invoked.

Some of the constructor and defaults tests are still good to have
around so those have been migrated over; however something about
the organization of the tests have created some confusion.  For
example, we never actually test we can't call the `initialize`
twice.

Hopefully, the new structure will help make it clear what is being
tested and should be tested.