Updated PoolTokenProxy tests
As with mAPT proxy tests, a big chunk was checking only OZ code
paths.  Those were removed and the ones remaining are checking
the initialization logic.  These were cleaned-up and extended.

Since we do have a V2 contract, there was no actual need for a
test upgrade contract, so that was deleted.