Adjust extra storage gap and correct comments
Previous comments and code were confused with regard to which
contract's slot positioning was relevant.  Shifting storage
down on the logic contract means the relevant slot on the proxy
storage will be increased by the extra storage gap.

Some manual testing was done to check that making the `_notEntered`
slot land in an OZ gap area would revert on proxy upgrade.  This
added to the confidence I have that this really does work for
getting around the re-entrancy guard (and that it is the cause of
the reverts).