Handle minting in pre-transfer hook
This interaction between `_mint` and the pre-transfer hook
was noted when attempting to initialize the logic contract during
testnet deployment, per our usual best practice for upgrades.

Since the logic contract `initialize` mints the deployer the total
supply, the blocked mint causes a "LOCKED_BALANCE" revert.

This is the only transfer-like situation that was not tested and is
only ever used in the `initialize` function.  This function is not
something you would ever be allowed to call again in production,
as it has been called already during original deployment.

Nonetheless we have changed the pre-transfer hook to allow the
check to pass when `from` is the zero address.  This will allow
minting as a feature in any future upgrade.